requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloud-sql-python-connector==1.2.1
cryptography==40.0.1
distlib==0.3.6
filelock==3.11.0
Flask==2.2.3
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
google-auth==2.17.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
multidict==6.0.4
pg8000==1.29.4
pipenv==2023.3.20
pkgutil-resolve-name==1.3.10
platformdirs==3.2.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
rsa==4.9
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.9
typing-extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
virtualenv-clone==0.5.7
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
